saving token in localstorage

+++ src/helpers/CoordinatesHelper.ts
@@ -3,7 +3,7 @@ export const splitCoordinates = (
 ): { latitude: number; longitude: number } => {
   const splitsedCoordinates = coordinates.split(",");
   return {
-    latitude: Number.parseInt(splitsedCoordinates[0]),
-    longitude: Number.parseInt(splitsedCoordinates[1]),
+    latitude: Number.parseFloat(splitsedCoordinates[0]),
+    longitude: Number.parseFloat(splitsedCoordinates[1]),
   };
 };
